refactor: move team-member.jsx to team-member folder
